Move all registry and registry store files to `infra/registry/`
