+++ sdk/python/feast/infra/materialization/lambda/lambda_engine.py
@@ -22,7 +22,7 @@ from feast.infra.materialization.batch_materialization_engine import (
 )
 from feast.infra.offline_stores.offline_store import OfflineStore
 from feast.infra.online_stores.online_store import OnlineStore
-from feast.registry import BaseRegistry
+from feast.infra.registry.registry import BaseRegistry
 from feast.repo_config import FeastConfigBaseModel, RepoConfig
 from feast.stream_feature_view import StreamFeatureView
 from feast.utils import _get_column_names
