+++ sdk/python/feast/infra/materialization/local_engine.py
@@ -9,10 +9,10 @@ from feast.entity import Entity
 from feast.feature_view import FeatureView
 from feast.infra.offline_stores.offline_store import OfflineStore
 from feast.infra.online_stores.online_store import OnlineStore
+from feast.infra.registry.registry import BaseRegistry
 from feast.repo_config import FeastConfigBaseModel, RepoConfig
 from feast.stream_feature_view import StreamFeatureView
 
-from ...registry import BaseRegistry
 from ...utils import (
     _convert_arrow_to_proto,
     _get_column_names,
