+++ sdk/python/feast/infra/offline_stores/snowflake.py
@@ -40,13 +40,13 @@ from feast.infra.offline_stores.snowflake_source import (
     SnowflakeLoggingDestination,
     SnowflakeSource,
 )
+from feast.infra.registry.registry import BaseRegistry
 from feast.infra.utils.snowflake_utils import (
     execute_snowflake_statement,
     get_snowflake_conn,
     write_pandas,
     write_parquet,
 )
-from feast.registry import BaseRegistry
 from feast.repo_config import FeastConfigBaseModel, RepoConfig
 from feast.saved_dataset import SavedDatasetStorage
 from feast.usage import log_exceptions_and_usage
