+++ sdk/python/feast/infra/registry/contrib/postgres/postgres_registry_store.py
@@ -3,10 +3,10 @@ from typing import Optional
 import psycopg2
 from psycopg2 import sql
 
+from feast.infra.registry.registry_store import RegistryStore
 from feast.infra.utils.postgres.connection_utils import _get_conn
 from feast.infra.utils.postgres.postgres_config import PostgreSQLConfig
 from feast.protos.feast.core.Registry_pb2 import Registry as RegistryProto
-from feast.registry_store import RegistryStore
 from feast.repo_config import RegistryConfig
