+++ sdk/python/tests/integration/e2e/test_usage_e2e.py
@@ -70,10 +70,10 @@ def test_usage_on(dummy_exporter, enabling_toggle):
 
         assert len(dummy_exporter) == 3
         assert {
-            "entrypoint": "feast.infra.local.FileRegistryStore.get_registry_proto"
+            "entrypoint": "feast.infra.registry.file.FileRegistryStore.get_registry_proto"
         }.items() <= dummy_exporter[0].items()
         assert {
-            "entrypoint": "feast.infra.local.FileRegistryStore.update_registry_proto"
+            "entrypoint": "feast.infra.registry.file.FileRegistryStore.update_registry_proto"
         }.items() <= dummy_exporter[1].items()
         assert {
             "entrypoint": "feast.feature_store.FeatureStore.apply"
