+++ sdk/python/tests/integration/registration/test_registry.py
@@ -23,7 +23,7 @@ from feast.data_format import ParquetFormat
 from feast.entity import Entity
 from feast.feature_view import FeatureView
 from feast.field import Field
-from feast.registry import Registry
+from feast.infra.registry.registry import Registry
 from feast.repo_config import RegistryConfig
 from feast.types import Array, Bytes, Int64, String
 from tests.utils.e2e_test_validation import validate_registry_data_source_apply
